src/third_party/blink/web_tests/virtual/web-components-v0-disabled/html/dialog/modal-dialog-distributed-child-is-not-inert-expected.txt

CONSOLE ERROR: line 28: Uncaught TypeError: document.querySelector(...).createShadowRoot is not a function
Tests that nodes transposed into the dialog are not inert. The test passes if you can click the button.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL Unexpected error: Uncaught TypeError: document.querySelector(...).createShadowRoot is not a function
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE

Click me
